libc: add page number for c99 standard

+++ libc/inc/stdio.h
@@ -1,11 +1,7 @@
 #pragma once
-#include <stdint.h>
 #include <stdarg.h>
-#include <stdbool.h>
-#include <stdatomic.h>
-#include <stddef.h>
-#include <stdfix.h>
-#include <stdnoreturn.h>
+
+// page 274 of https://www.open-std.org/jtc1/sc22/wg14/www/docs/n1256.pdf
 
 // alias the underscore names of the printf standalone library
 #define printf_ printf
@@ -20,8 +16,8 @@ int printf_(const char *format, ...);
 int vprintf_(const char *format, va_list arg);
 int sprintf_(char *s, const char *format, ...);
 int vsprintf_(char *s, const char *format, va_list arg);
-int snprintf_(char *s, size_t count, const char *format, ...);
-int vsnprintf_(char *s, size_t count, const char *format, va_list arg);
+int snprintf_(char *s, unsigned long count, const char *format, ...);
+int vsnprintf_(char *s, unsigned long count, const char *format, va_list arg);
 int fctprintf(void (*out)(char c, void *extra_arg), void *extra_arg, const char *format, ...);
 int vfctprintf(void (*out)(char c, void *extra_arg), void *extra_arg, const char *format, va_list arg);
